Add SearchResultContainer component
